chore: moved components to common
