+++ app/src/main/java/com/contextu/al/announcement/AnnouncementComponent.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import com.contextu.al.common.ContainerComponent
 import com.contextu.al.common.ImageComponent
-import com.contextu.al.components.AppTextView
-import com.contextu.al.components.ButtonComponent
+import com.contextu.al.common.AppTextView
+import com.contextu.al.common.ButtonComponent
 import com.contextu.al.model.customguide.GuidePayload
 import com.contextu.al.model.toBoxAlignment
